requirements.txt
asgiref==3.8.1
boto3==1.37.18
botocore==1.37.18
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
cssbeautifier==1.15.2
dj-database-url==2.3.0
Django==5.1.6
django-allauth==65.3.1
django-storages==1.14.5
django-widget-tweaks==1.5.0
djlint==1.36.4
EditorConfig==0.17.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
jsbeautifier==1.15.2
json5==0.10.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
